refactor: improve pokeDetails component
