+++ src/components/PokeDetails.vue
@@ -1,52 +1,10 @@
 <script setup>
-import { ref, reactive } from "vue";
+import useGetPokemonDetails from "../hooks/useGetPokemonDetails.js";
 
 const props = defineProps({
    pokeName: String,
 });
-
-const poke = reactive({
-   name: "",
-   description: "",
-   avatar: "",
-   types: [],
-   abilities: [],
-   height: 0,
-   weight: 0,
-   stats: [],
-   weakness: [],
-});
-const loading = ref(true);
-const single = `https://pokeapi.co/api/v2/pokemon/${props.pokeName}/`;
-const specie = `https://pokeapi.co/api/v2/pokemon-species/${props.pokeName}/`;
-
-const getSingleDetails = async (url) => {
-   const res = await fetch(url);
-   const data = await res.json();
-   poke.name = data.name;
-   poke.avatar = data.sprites.other.dream_world.front_default;
-   poke.types = data.types.map((el) => el.type.name);
-   poke.abilities = data.abilities.map((el) => el.ability.name);
-   poke.height = data.height;
-   poke.weight = data.weight;
-   poke.stats = data.stats.map((el) => ({
-      type: el.stat.name,
-      value: el.base_stat,
-   }));
-};
-const getSpecieDetails = async (url) => {
-   const res = await fetch(url);
-   const data = await res.json();
-   poke.description = data.flavor_text_entries[0].flavor_text;
-};
-
-const getAllData = () => {
-   getSingleDetails(single);
-   getSpecieDetails(specie);
-   loading.value = false;
-};
-
-getAllData();
+const { poke, loading } = useGetPokemonDetails(props.pokeName);
 </script>
 
 <template>
